asserts: move type annotations to an asserts.pyi stubs file
Let's clearly separate the typing and implementation, instead of using a
large `TYPE_CHECKING` block. Stubs also have more flexible rules, like
not needed an implementation function signature for overloads.
